Think I found the solidity error

+++ HashgraphVisualizerSDK/Voting.sol
@@ -23,7 +23,7 @@ contract Voting {
     return votesReceived[candidate];
   }
 
-  function vote(uint8[] votes, uint voterId) public {
+  function vote(uint8[] votes, uint8 voterId) public {
     for (uint i = 0; i < candidateList.length; i++) {
       assert(voterBalance[voterId] >= votes[i]);
       voterBalance[voterId] -= votes[i];
